PLANET-7625: Move scripts from twig templates
- Make code clearer
