+++ src/Loader.php
@@ -84,7 +84,7 @@ final class Loader
             GravityFormsExtensions::class,
             BlockSettings::class,
             MediaReplacer::class,
-            TwigScriptsEnqueuer::class,
+            EnqueueController::class,
         ];
 
         if (is_admin()) {
